+++ backend/windmill-common/src/scripts.rs
@@ -25,21 +25,27 @@ use crate::utils::StripPath;
 )]
 #[serde(rename_all(serialize = "lowercase", deserialize = "lowercase"))]
 pub enum ScriptLang {
+    Nativets,
     Deno,
     Python3,
     Go,
     Bash,
     Postgresql,
+    Bun,
+    Mysql,
 }
 
 impl ScriptLang {
     pub fn as_str(&self) -> &'static str {
         match self {
+            ScriptLang::Bun => "bun",
+            ScriptLang::Nativets => "nativets",
             ScriptLang::Deno => "deno",
             ScriptLang::Python3 => "python3",
             ScriptLang::Go => "go",
             ScriptLang::Bash => "bash",
             ScriptLang::Postgresql => "postgresql",
+            ScriptLang::Mysql => "mysql",
         }
     }
 }