+++ backend/windmill-worker/src/worker_flow.rs
@@ -1693,7 +1693,6 @@ async fn compute_next_flow_transform(
     };
     match &module.value {
         FlowModuleValue::Identity => trivial_next_job(JobPayload::Identity),
-        FlowModuleValue::Graphql => trivial_next_job(JobPayload::Graphql),
         FlowModuleValue::Http => trivial_next_job(JobPayload::Http),
         FlowModuleValue::Flow { path, .. } => {
             let payload = JobPayload::Flow(path.to_string());