+++ cli/main.ts
@@ -28,7 +28,7 @@ addEventListener("error", (event) => {
   }
 });
 
-export const VERSION = "v1.124.0";
+export const VERSION = "v1.125.1";
 
 let command: any = new Command()
   .name("wmill")