+++ deno-client/mod.ts
@@ -21,13 +21,6 @@ export {
   WorkspaceService,
 } from "./windmill-api/index.ts";
 
-// @ts-ignore: Otherwise BigInt is not supported for export
-BigInt.prototype.toJSON = function () {
-  return this.toString();
-};
-
-export { pgSql, pgClient } from "./pg.ts";
-
 export type Sql = string;
 export type Email = string;
 export type Base64 = string;