+++ frontend/src/lib/components/apps/components/helpers/RunnableComponent.svelte
@@ -181,6 +181,8 @@
 				await setResult(r, undefined)
 
 				$state = $state
+				const job = generateNextFrontendJobId()
+				$jobs = [{ job, component: id, result: r }, ...$jobs]
 			} catch (e) {
 				sendUserToast(`Error running frontend script ${id}: ` + e.message, true)
 
@@ -298,7 +300,10 @@
 				$worldStore.newOutput(id, 'raw', res)
 				res = await eval_like(
 					transformer.content,
-					computeGlobalContext($worldStore, { result: res }),
+					computeGlobalContext(
+						$worldStore,
+						iterContext ? { iter: $iterContext, result: res } : { result: res }
+					),
 					false,
 					$state,
 					$mode == 'dnd',
@@ -360,6 +365,8 @@
 		undefined
 
 	onMount(() => {
+		console.log('create', id)
+
 		cancellableRun = (inlineScript?: InlineScript) => {
 			let rejectCb: (err: Error) => void
 			let p: Partial<CancelablePromise<void>> = new Promise<void>((resolve, reject) => {