+++ frontend/src/lib/components/flows/common/FlowCardHeader.svelte
@@ -35,7 +35,7 @@
 				{#if flowModule.value.type === 'identity'}
 					<span class="font-bold text-xs">Identity (input copied to output)</span>
 				{:else if flowModule?.value.type === 'rawscript'}
-					<div class="w-4 mx-0.5">
+					<div class="w-8 mx-0.5">
 						<LanguageIcon lang={flowModule.value.language} class="w-4 h-4" />
 					</div>
 					<input bind:value={flowModule.summary} placeholder={'Summary'} class="w-full grow" />